requirements.txt
Flask==2.3.3
Jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.0
unittest-xml-reporting==3.0.6
